pyproject.toml
[project]
name = "azure-devops-mcp"
version = "0.1.0"
description = "MCP server exposing 41 Azure DevOps tools to AI assistants — pipelines, repos, PRs, boards, releases, tests, wikis"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.10"
keywords = ["mcp", "azure-devops", "ai", "copilot", "claude", "model-context-protocol"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Topic :: Software Development :: Libraries",
    "Intended Audience :: Developers",
]
dependencies = [
    "mcp[cli]>=1.0.0",
    "msal>=1.20.0",
    "python-dotenv>=1.0.0",
    "httpx>=0.27.0",
]

[project.optional-dependencies]
managed-identity = ["azure-identity>=1.15.0"]
dev = ["pytest>=7.0.0", "pytest-cov>=4.0.0"]

[project.urls]
Homepage = "https://github.com/imsaurav/azure-devops-mcp"
Repository = "https://github.com/imsaurav/azure-devops-mcp"
Issues = "https://github.com/imsaurav/azure-devops-mcp/issues"

[project.scripts]
azure-devops-mcp = "azure_devops_mcp.__main__:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/azure_devops_mcp"]
